+++ Magic/CHANGELOG.md
@@ -10,6 +10,7 @@
  - Fix error in /spells
  - Add netherrite armor and tools to wearable and melee material lists
  - Armor can now be made to be worn in a slot other than the vanilla slot it's supposed to go in (see: hardhat)
+ - Undroppable wands can not be used as crafting ingredients
 
 # 7.10.4